removed h2 from contract-definition
