+++ extensions/sql/contract-definition-store-sql/build.gradle.kts
@@ -17,7 +17,6 @@ plugins {
     `maven-publish`
 }
 
-val h2Version: String by project
 val assertj: String by project
 val postgresVersion: String by project
 
@@ -32,7 +31,6 @@ dependencies {
 
     testImplementation(project(":extensions:junit"))
     testImplementation(project(":core:base"))
-    testImplementation("com.h2database:h2:${h2Version}")
     testImplementation("org.assertj:assertj-core:${assertj}")
     testImplementation(testFixtures(project(":common:util")))
     testImplementation("org.postgresql:postgresql:${postgresVersion}")
